Change alert information to be stored in session rather than a class
As a user will only be seeing one alert at a time, the alert information can be stored in the session, allowing it to be accessed at any time.

The function set_alert() in web_pages.py allows you to change these settings in what I beleive to be a more pythonic way.